:white_check_mark: Add unit tests for serializers
